requirements.txt
groq>=0.4.0
datasets>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
seaborn>=0.12.0
google-genai>=1.0.0
google-auth>=2.0.0
